requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
opencv-python>=4.8.0
huggingface_hub>=0.20.0
datasets>=2.16.0
grad-cam>=1.4.8
